tried to make the changes as requested
