+++ app/src/main/AndroidManifest.xml
@@ -18,15 +18,8 @@
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE_DATA_SYNC" />
     <uses-permission android:name="android.permission.USE_BIOMETRIC" />
     <uses-permission android:name="android.permission.REQUEST_IGNORE_BATTERY_OPTIMIZATIONS" /> <!-- We can use this directly as CS3 is not on Play Store -->
-
-    <!--    start-->
-    <uses-permission android:name="com.android.providers.tv.permission.READ_EPG_DATA" />
-    <uses-permission android:name="com.android.providers.tv.permission.WRITE_EPG_DATA" />
-    <uses-permission android:name="com.android.providers.tv.permission.WRITE_PREVIEW_DATA" />
-    <uses-permission android:name="com.android.providers.tv.permission.ACCESS_ALL_EPG_DATA" />
-    <uses-permission android:name="android.permission.READ_TV_LISTINGS" />
-    <!--    end-->
-
+    <uses-permission android:name="com.android.providers.tv.permission.READ_EPG_DATA" />  <!-- We can use to read the tv channel list -->
+    <uses-permission android:name="com.android.providers.tv.permission.WRITE_EPG_DATA" />  <!-- We can use this to write programs into the channel -->
     <!-- Required for OpenInAppAction and getting arbitrary Aniyomi packages  -->
     <uses-permission
         android:name="android.permission.QUERY_ALL_PACKAGES"
