+++ app/src/main/java/com/lagradost/cloudstream3/utils/TvChannelUtils.kt
@@ -1,0 +1,133 @@
+package com.lagradost.cloudstream3.utils
+
+import android.content.ContentUris
+import android.content.Context
+import android.content.Intent
+import android.net.Uri
+import android.util.Log
+import androidx.tvprovider.media.tv.PreviewProgram
+import androidx.tvprovider.media.tv.TvContractCompat
+import com.lagradost.cloudstream3.base64Encode
+import com.lagradost.cloudstream3.syncproviders.AccountManager.Companion.APP_STRING_SHARE
+import com.lagradost.cloudstream3.MainActivity
+import com.lagradost.cloudstream3.mapper
+import com.lagradost.cloudstream3.syncproviders.AccountManager.Companion.APP_STRING_SHARE
+import com.lagradost.cloudstream3.ui.home.HomeFragment
+import com.lagradost.cloudstream3.utils.DataStore.getKey
+import com.lagradost.cloudstream3.utils.DataStore.setKey
+
+const val PROGRAM_ID_LIST_KEY = "persistent_program_ids"
+
+class TvChannelUtils {
+    fun Context.saveProgramId(programId: Long) {
+        val existing: List<Long> = getKey(PROGRAM_ID_LIST_KEY) ?: emptyList()
+        val updated = (existing + programId).distinct()
+        setKey(PROGRAM_ID_LIST_KEY, updated)
+    }
+    fun Context.getStoredProgramIds(): List<Long> {
+        return getKey(PROGRAM_ID_LIST_KEY) ?: emptyList()
+    }
+    fun Context.removeProgramId(programId: Long) {
+        val existing: List<Long> = getKey(PROGRAM_ID_LIST_KEY) ?: emptyList()
+        val updated = existing.filter { it != programId }
+        setKey(PROGRAM_ID_LIST_KEY, updated)
+    }
+
+
+    fun getChannelId(context: Context, channelName: String): Long? {
+        return try {
+            context.contentResolver.query(
+                TvContractCompat.Channels.CONTENT_URI,
+                arrayOf(
+                    TvContractCompat.Channels._ID,
+                    TvContractCompat.Channels.COLUMN_DISPLAY_NAME
+                ),
+                null,
+                null,
+                null
+            )?.use { cursor ->
+                while (cursor.moveToNext()) {
+                    val id = cursor.getLong(
+                        cursor.getColumnIndexOrThrow(TvContractCompat.Channels._ID)
+                    )
+                    val name = cursor.getString(
+                        cursor.getColumnIndexOrThrow(TvContractCompat.Channels.COLUMN_DISPLAY_NAME)
+                    )
+                    if (name == channelName) return id
+                }
+                null
+            }
+        } catch (e: Exception) {
+            Log.e("TvChannelUtils", "Query failed: ${e.message}", e)
+            null
+        }
+    }
+
+    /** Insert programs into a channel */
+
+    fun addPrograms(context: Context, channelId: Long, items: List<HomeFragment.SearchResponseImpl>) {
+        for (item in items) {
+            try {
+                val nameBase64 = base64Encode(item.apiName.toByteArray(Charsets.UTF_8))
+                val urlBase64 = base64Encode(item.url.toByteArray(Charsets.UTF_8))
+                val csshareUri = "$APP_STRING_SHARE:$nameBase64?$urlBase64"
+                val intent = Intent(Intent.ACTION_VIEW, Uri.parse(csshareUri))
+                val builder = PreviewProgram.Builder()
+                    .setChannelId(channelId)
+                    .setTitle(item.name)
+                    .setDescription(item.apiName)
+                    .setContentId(item.url)
+                    .setType(TvContractCompat.PreviewPrograms.TYPE_CLIP)
+                    .setIntentUri(Uri.parse(csshareUri))
+                    .setPosterArtAspectRatio(TvContractCompat.PreviewPrograms.ASPECT_RATIO_2_3)
+
+                // Validate poster URL before setting
+                val posterUrl = item.posterUrl
+                if (!posterUrl.isNullOrBlank() && posterUrl.startsWith("http")) {
+                    builder.setPosterArtUri(Uri.parse(posterUrl))
+                } else {
+                    Log.w("TvChannelUtils", "Missing or invalid poster for ${item.name}")
+                }
+
+                val program = builder.build()
+
+                val uri = context.contentResolver.insert(
+                    TvContractCompat.PreviewPrograms.CONTENT_URI,
+                    program.toContentValues()
+                )
+
+                if (uri != null) {
+                    val programId = ContentUris.parseId(uri)
+                    context.saveProgramId(programId)
+                    Log.d("TvChannelUtils", "Inserted program ${item.name}, ID=$programId")
+                } else {
+                    Log.e("TvChannelUtils", "Insert failed for ${item.name}")
+                }
+
+            } catch (error: Exception) {
+                Log.e("TvChannelUtils", "Error inserting ${item.name}: $error")
+            }
+        }
+    }
+
+    fun deleteStoredPrograms(context: Context) {
+        val programIds = context.getStoredProgramIds()
+
+        for (id in programIds) {
+            val uri = ContentUris.withAppendedId(TvContractCompat.PreviewPrograms.CONTENT_URI, id)
+            try {
+                val rowsDeleted = context.contentResolver.delete(uri, null, null)
+                if (rowsDeleted > 0) {
+                    context.removeProgramId(id) // Remove from persistent list
+                } else {
+                    Log.w("ProgramDelete", "No program found for ID: $id")
+                }
+            } catch (e: Exception) {
+                Log.e("ProgramDelete", "Failed to delete program ID: $id", e)
+            }
+        }
+
+        Log.d("ProgramDelete", "Finished deleting stored programs")
+    }
+
+}
